Move karma inflation cache into lib
