+++ packages/lesswrong/lib/collections/posts/karmaInflation.ts
